+++ src/socket.c
@@ -806,9 +806,10 @@ int ssh_socket_set_blocking(socket_t fd)
  * @bug It only tries connecting to one of the available AI's
  * which is problematic for hosts having DNS fail-over.
  */
-
-int
-ssh_socket_connect(ssh_socket s, const char *host, int port, const char *bind_addr)
+int ssh_socket_connect(ssh_socket s,
+                       const char *host,
+                       uint16_t port,
+                       const char *bind_addr)
 {
     socket_t fd;
